Enable clustering for gene set heatmaps
List any queried gene set profile in the heatmap menu, displaying only
the clustering button as it's queried via the query page and the ‘Add
Genes’ and ‘Remove Heatmap’ buttons do not apply to it.

Collapse expansion tracks before clustering a heatmap, as these are
generally not of the same data type as the main tracks of the heatmap,
and including them when clustering would not make sense.